Refactor game logic use
Features:
- Added GameEngine class containing game logic method
- Added Observation class containing the observation of the board used by the DQN algo
- Added PlayerSelector class
- Added InputHandler class to manage I/O

+++ src/GameEngine/InputHandler.cpp
@@ -61,7 +61,7 @@ int InputHandler::readMove()
 
     cin >> moveStr;
 
-    for (int i = 0; i < ACTIONS.size(); i++) {
+    for (size_t i = 0; i < ACTIONS.size(); i++) {
         if (ACTIONS[i] == moveStr) {
             move = i;
             break;
